fix: refactor code into a separate directory
